Integrate AI model for pitch detection using ONNX Runtime
Refactor AIModelLoader and PitchCorrectionEngine to support ONNX Runtime, including conditional inclusion of ONNX headers and adjustments to data processing loops and type casting for improved performance and compatibility.

+++ .ccls-cache/@home@runner@workspace/Source@AIModelLoader.cpp
@@ -5,6 +5,7 @@
 #include <future>
 #include <memory>
 #include <juce_core/juce_core.h>
+#include <juce_audio_utils/juce_audio_utils.h>
 
 // PIMPL implementation to hide complex AI model details
 struct AIModelLoader::Impl
@@ -36,7 +37,7 @@ struct AIModelLoader::Impl
                 float correlation = 0.0f;
                 int count = 0;
                 
-                for (int i = 0; i < static_cast<int>(audio.size()) - period; ++i)
+                for (size_t i = 0; i < audio.size() - static_cast<size_t>(period); ++i)
                 {
                     correlation += audio[i] * audio[i + period];
                     count++;
